requirement.txt
discord
python-dotenv
pynacl
pytube
ffmpeg-python
pyttsx3
SpeechRecognition
pyaudio
gtts